requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
lxml>=4.9.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
python-dotenv>=1.0.0